Updated pause menu to use InputMaster system (should recieve inputs based on current control scheme); Addressed an issue with the save system sometimes causing a crash when starting a new game (was previously deleting a freshly generated save file and refusing to launch the maintenance scene

+++ Assets/Scripts/Menus/MainMenuController.cs
@@ -145,7 +145,7 @@ public class MainMenuController : MonoBehaviour
         //    //If a save file is found, display save file confirmation panel
         //}
 
-        if (Directory.Exists(System.IO.Path.Combine(Application.persistentDataPath, "Save")))
+        if (Directory.Exists(System.IO.Path.Combine(Application.persistentDataPath, "Save")) && SaveDataController.instance.saveData.currentScene != "")
         {
             print("Deleting old save file");
             Directory.Delete(System.IO.Path.Combine(Application.persistentDataPath, "Save"), true);
